K, I couldn't help myself, I can't leave shit unfinished. The Symbol table now functions properly in all senses of the word, and output produces a very neatly formatted string ready to be printed to the screen with a list of all the data that is required (as far as I know at this point).
NOTE: IF YOU AREN'T USING WINDOWS, THIS WILL NOT PRINT OUT PRETTILY. If I recall correctly, almost everyone is on or has access to a windows machine, so I just went ahead and used \n for new lines and \t for tabs. I honestly don't remember what the difference is for Linux based OS's, or OSX, but I'm pretty sure there is one.
If we decide to change the format, it will only take 30 seconds to do so, so no worries @ that. Anyways, I tested it, it looks like it's working, and I'm happy for the night. Feel free to test on your own, and let me know if it needs more functionality.

Jeffman OUT!